requirements.txt
fastapi==0.116.1
uvicorn==0.30.6
pandas==2.2.2
numpy==1.26.4
scikit-learn==1.5.1
lightgbm==4.5.0
sentence-transformers==3.0.1
rank-bm25==0.2.2
faiss-cpu==1.8.0.post1
python-multipart==0.0.9
pytest==8.3.4
